Refactor Psylean app architecture
