+++ Psylean/Core/Components/ErrorView.swift
@@ -1,0 +1,38 @@
+//
+//  ErrorView.swift
+//  Psylean
+//
+//  Created by Rudrank Riyam on 26/06/25.
+//
+
+import SwiftUI
+
+struct ErrorView: View {
+    let error: Error
+    let retry: () -> Void
+    
+    var body: some View {
+        VStack(spacing: 16) {
+            Image(systemName: "exclamationmark.triangle")
+                .font(.largeTitle)
+                .foregroundColor(.red)
+            
+            Text("Failed to fetch Pokemon data")
+                .font(.headline)
+            
+            Text(error.localizedDescription)
+                .font(.caption)
+                .foregroundStyle(.secondary)
+                .multilineTextAlignment(.center)
+            
+            Button("Retry") {
+                retry()
+            }
+            .buttonStyle(.borderedProminent)
+        }
+        .padding()
+        #if os(iOS) || os(macOS)
+        .glassEffect(.regular, in: .rect(cornerRadius: 16))
+        #endif
+    }
+}
