+++ Psylean/Features/Analysis/Models/PokemonAnalysis.swift
